+++ src/extractors/emojiextractor.py
@@ -8,59 +8,65 @@ from lxml import etree
 from lxml.etree import XPath
 from tqdm import tqdm
 
+from .blockfactory import Block
 from .characterfactory import Character
 from .extractor import Extractor
 
 
 class EmojiExtractor(Extractor):
-    __all_emojis: Dict[str, Character]
+    __all_blocks: List[Block]
+    __annotations: Dict[str, List[str]]
     __base_emojis: List[str]
 
     def __init__(self):
-        self.__all_emojis = {}
+        self.__annotations = {}
+        self.__all_blocks = []
         self.__base_emojis = []
 
     def __fetch_data(self) -> None:
         with tqdm(total=3) as progress:
-            progress.set_description("Downloading list of all emojis")
-            self.__all_emojis = self.__fetch_emoji_list()
-            progress.update(1)
-
             progress.set_description("Downloading annotations")
             self.__fetch_annotations()
+            progress.update(1)
+
+            progress.set_description("Downloading list of all emojis")
+            self.__fetch_emoji_list()
             progress.update(1)
 
             progress.set_description("Downloading list of human emojis")
             self.__base_emojis = self.__fetch_base_emojis()
             progress.update(1)
 
-    def __fetch_emoji_list(self) -> Dict[str, Character]:
+    def __fetch_emoji_list(self) -> None:
         data: requests.Response = requests.get(
             "https://unicode.org/emoji/charts-15.0/full-emoji-list.html", timeout=120
         )
 
         html = BeautifulSoup(data.text, "lxml")
 
-        emojis: Dict[str, Character] = {}
+        current_title = None
+        current_emojis: List[Character] = []
         for row in html.find("table").find_all("tr"):
-            if not row.th:
+            if row.th and "bighead" in row.th["class"]:
+                if current_title:
+                    self.__all_blocks.append(Block(current_title, current_emojis))
+                current_title = row.a.string
+                current_emojis = []
+            elif not row.th:
                 emoji = row.find("td", {"class": "chars"}).string
                 description = row.find("td", {"class": "name"}).string.replace("⊛ ", "")
-                emojis[emoji] = Character(emoji, description, "L")
+                current_emojis.append(Character(emoji, description, "L", self.__annotations.get(emoji, None)))
 
-        return emojis
+        self.__all_blocks.append(Block(current_title, current_emojis))
 
-    def __fetch_annotations(self):
+    def __fetch_annotations(self) -> None:
         data: requests.Response = requests.get(
             "https://raw.githubusercontent.com/unicode-org/cldr/latest/common/annotations/en.xml", timeout=60
         )
 
         xpath = XPath('./annotations/annotation[not(@type="tts")]')
         for element in xpath(etree.fromstring(data.content)):
-            try:
-                self.__all_emojis[element.get("cp")].add_descriptions(element.text.split(" | "))
-            except KeyError:
-                pass
+            self.__annotations[element.get("cp")] = element.text.split(" | ")
 
     def __fetch_base_emojis(self) -> List[str]:
         data: requests.Response = requests.get("https://unicode.org/Public/15.0.0/ucd/emoji/emoji-data.txt", timeout=60)
@@ -90,13 +96,16 @@ class EmojiExtractor(Extractor):
         return "".join(chr(int(character, 16)) for character in string.split(" "))
 
     def __write_symbol_file(self, target: Path) -> None:
-        with (target / "emojis.csv").open("w") as symbol_file:
-            for entry in self.__compile_entries(self.__all_emojis):
-                symbol_file.write(entry + "\n")
+        for block in self.__all_blocks:
+            with (target / f"emojis_{block.name.lower().replace(' ', '').replace('&', '_')}.csv").open(
+                "w"
+            ) as symbol_file:
+                for entry in self.__compile_entries(block):
+                    symbol_file.write(entry + "\n")
 
-    def __compile_entries(self, emojis: Dict[str, Character]) -> List[str]:
+    def __compile_entries(self, block: Block) -> List[str]:
         annotated_emojis = []
-        for emoji in emojis.values():
+        for emoji in block.characters:
             entry = f"{emoji.char} {html.escape(emoji.name)}"
             if emoji.descriptions:
                 entry += f" <small>({html.escape(', '.join(emoji.descriptions))})</small>"