+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Changed
 - `rofimoji` uses [Poetry](https://python-poetry.org/) for package management. With this comes the requirement for Python 3.8.
 - `--files` supports globbing, so you can ask for `--files "linear_b*"`
+- The emoji file is split into several files for one category each. (#126)
 
 ## Added
 - You can now select your desired actions through a separate menu.